* 'develop' of github.com:genropy/genropy: (1031 commits)
  set GNR_DAEMON_HOST to local interface when deploying fullstack
  support multiple environment secret to be passed, which will be joined by k8s.
  output build configuration file in indented format for readability, much better than a one liner json
  EXTERNAL_HOST always forced to https
  new version bump
  Fixes malformed publish action and improves readability (#216)
  introducing action/cache@v3
  added missing depedency "unidecode", not detected as it was a runtime import.
  fixed metadata name to align with traefik ingress rules
  for site daemon, use cli options for host if provided, taking precedence over configuration file.
  Picker support for store_kwargs (#215)
  provide sitename and instance name to daemon
  removed a logger warning
  when running a site daemon, the port options on CLI has precedence over default.xml for siteconfig
  Generated Persistent volume generation for site folder
  correctly invoke site daemon, need to use -n to provide the site when booting up the stack
  removed unused constants, formatting cleanup
  added procps and iproute2 to base genropy image, needed to debug connection issues
  Improves endpoint URL resolution logic
  Fixes file extension parsing for URLs in DOM handler
  ...

# Conflicts:
#	projects/gnrcore/packages/email/localization.xml
#	projects/gnrcore/packages/sys/webpages/localizationeditor.py
#	projects/gnrcore/packages/test/webpages/tools/ckeditor.py

+++ projects/gnrcore/packages/docu/resources/tables/handbook/action/export_to_sphinx.py
@@ -3,13 +3,12 @@
 # test_special_action.py
 # Created by Francesco Porcari on 2010-07-02.
 # Copyright (c) 2011 Softwell. All rights reserved.
-from gnr.web.batch.btcbase import BaseResourceBatch
-from gnr.app.gnrlocalization import AppLocalizer
-from gnr.core.gnrbag import Bag
+
 from json import dumps
 from datetime import datetime
 import re
 import sys
+import time
 
 if sys.version_info[0] == 3:
     from urllib.request import urlopen
@@ -18,6 +17,14 @@ else:
     # But note that this might need an update when Python 4
     # might be around one day
     from urllib import urlopen
+
+from sphinx.cmd.build import main as sphinx_build_main
+import boto3
+
+from gnr.web.batch.btcbase import BaseResourceBatch
+from gnr.app.gnrlocalization import AppLocalizer
+from gnr.core.gnrbag import Bag
+from gnr.app import pkglog as logger
 
 caption = 'Export to sphinx'
 description = 'Export to sphinx'
@@ -30,11 +37,11 @@ class Main(BaseResourceBatch):
     batch_title =  'Export to sphinx'
     batch_cancellable = False
     batch_delay = 0.5
-    batch_steps = 'prepareRstDocs,buildHtmlDocs'
+    batch_steps = 'prepareConfFile,prepareRstDocs,buildHtmlDocs'
 
     def pre_process(self):
         self.handbook_id = self.batch_parameters['extra_parameters']['handbook_id']
-        self.handbook_record = self.tblobj.record(self.handbook_id).output('bag')
+        self.handbook_record = self.tblobj.record(self.handbook_id, virtual_columns='$sphinx_path').output('bag')
         self.doctable =self.db.table('docu.documentation')
         self.doc_data = self.doctable.getHierarchicalData(root_id=self.handbook_record['docroot_id'], condition='$is_published IS TRUE')['root']['#0']
         #DP202208 Temporary node to build files, moved after creation to definitive folder
@@ -44,21 +51,12 @@ class Main(BaseResourceBatch):
         #DP202208 publishedDocNode node is where the final documentation will be published, at beginning of the process we erase former docs
         self.publishedDocNode = self.page.site.storageNode(self.handbook_record['sphinx_path'])
         self.publishedDocNode.delete()
-        confSn = self.sourceDirNode.child('conf.py')
-        self.page.site.storageNode('rsrc:pkg_docu','sphinx_env','default_conf.py').copy(self.page.site.storageNode(confSn))
-        theme = self.handbook_record['theme'] or 'sphinx_rtd_theme'
-        theme_path = self.page.site.storageNode('rsrc:pkg_docu','sphinx_env','themes').internal_path
-        html_baseurl = self.db.application.getPreference('.sphinx_baseurl',pkg='docu') or self.page.site.externalUrl('') + 'docs/' 
-        #DP202111 Default url set to /docs
-        self.handbook_url = html_baseurl + self.handbook_record['name'] + '/'
-        extra_conf = """html_theme = '%s'\nhtml_theme_path = ['%s/']\nhtml_baseurl='%s'\nsitemap_url_scheme = '%s/{link}'"""%(theme, theme_path, html_baseurl,self.handbook_record['name'])
-        with confSn.open('a') as confFile:
-            confFile.write(extra_conf)
+        self.html_baseurl = self.db.application.getPreference('.sphinx_baseurl',pkg='docu') or self.page.site.externalUrl('/_documentation/')
+        self.handbook_url = f"{self.html_baseurl}{self.handbook_record['name']}/"
+        self.enable_sitemap = self.db.application.getPreference('.enable_sitemap',pkg='docu')
         self.imagesDict = dict()
         self.imagesPath='_static/images'
         self.examplesPath='_static/_webpages'
-        self.customCssPath='_static/custom.css'
-        self.customJSPath='_static/custom.js'
         self.examples_root = None 
         self.examples_pars = Bag(self.handbook_record['examples_pars'])
         self.examples_mode = self.examples_pars['mode'] or 'iframe'
@@ -73,8 +71,41 @@ class Main(BaseResourceBatch):
         #DP202112 Check if there are active redirects
         if self.db.application.getPreference('.manage_redirects',pkg='docu'):
             self.redirect_pkeys = self.db.table('docu.redirect').query(where='$old_handbook_id=:h_id AND $is_active IS TRUE', 
-                        h_id=self.handbook_id).selection().output('pkeylist')
-
+                        h_id=self.handbook_id).selection().output('pkeylist')   
+            
+    def step_prepareConfFile(self):
+        "Prepare conf file"
+        confSn = self.sourceDirNode.child('conf.py')
+        self.page.site.storageNode('rsrc:pkg_docu','sphinx_env','default_conf.py').copy(self.page.site.storageNode(confSn))
+        theme = self.handbook_record['theme'] or 'sphinx_rtd_theme'
+        theme_path = self.page.site.storageNode('rsrc:pkg_docu','sphinx_env','themes').internal_path
+        handbooks_theme_pref = self.db.application.getPreference('.handbooks_theme',pkg='docu')
+        conf_lines = [
+            f"html_theme = '{theme}'",
+            f"html_theme_path = ['{theme_path}/']",
+            f"html_baseurl='{self.html_baseurl}'",
+        ]
+        if handbooks_theme_pref['logo']:
+            conf_lines.append(f"html_logo = '{self.page.site.externalUrl(handbooks_theme_pref['logo'])}'")
+            conf_lines.append("html_short_title = 'Handbook'")
+        if handbooks_theme_pref['copyright']:
+            conf_lines.append("show_copyright = True")
+            conf_lines.append("html_show_sphinx = False")
+            conf_lines.append(f"copyright = '{self.db.workdate.year} {handbooks_theme_pref['copyright']}'")
+        if handbooks_theme_pref['display_version']:
+            conf_lines.append("theme_display_version = True")
+            conf_lines.append(f"version = release = '{self.handbook_record['version']}'")
+        if handbooks_theme_pref['last_update']:
+            conf_lines.append("html_last_updated_fmt = '%d-%m-%Y'")
+        if handbooks_theme_pref['show_authors']:
+            conf_lines.append("show_authors = True")
+        if self.enable_sitemap: #DP Enabled extensions = ['sphinx_sitemap','sphinxext.opengraph']
+            conf_lines.append(f"sitemap_url_scheme = '{self.handbook_record['name']}/{{link}}'")
+        extra_conf = '\n'.join(conf_lines)
+        with confSn.open('a') as confFile:
+            confFile.write(extra_conf)
+            logger.info("Extra conf lines added to conf.py: %s" % extra_conf)
+            
     def step_prepareRstDocs(self):
         "Prepare Rst docs"
         if self.handbook_record['toc_roots']:
@@ -94,13 +125,16 @@ class Main(BaseResourceBatch):
 
         self.createFile(pathlist=[], name='index', title=self.handbook_record['title'], rst='', tocstring=tocstring)
         for k,v in self.imagesDict.items():
-            if self.batch_parameters.get('skip_images'): 
-                continue
             #DP202112 Useful for local debugging
-            source_url = self.page.externalUrl(v) if v.startswith('/') else v
+            source_url = self.page.site.externalUrl(v) if v.startswith('/') else v
             child = self.sourceDirNode.child(k)
             with child.open('wb') as f:
-                f.write(urlopen(source_url).read())
+                try:
+                    f.write(urlopen(source_url).read())
+                except:
+                    logger.debug("Missing file", source_url)
+                    continue
+                    
         for relpath,source in self.examplesDict.items():
             if not source:
                 continue
@@ -110,7 +144,7 @@ class Main(BaseResourceBatch):
     def step_buildHtmlDocs(self):
         "Build HTML docs"
         self.resultNode = self.sphinxNode.child('build')
-        ogp_image = self.page.externalUrl(self.handbook_record['ogp_image']) if self.handbook_record['ogp_image'] else None
+        ogp_image = self.page.site.externalUrl(self.handbook_record['ogp_image']) if self.handbook_record['ogp_image'] else None
         build_args = dict(project=self.handbook_record['title'],
                           version=self.handbook_record['version'],
                           author=self.handbook_record['author'],
@@ -125,13 +159,13 @@ class Main(BaseResourceBatch):
         for k,v in template_variables.items():
             if v:
                 args.extend(['-A', '%s=%s' % (k,v)])
-        customStyles = self.handbook_record['custom_styles'] or ''
-        customStyles = '%s\n%s' %(customStyles,self.defaultCssCustomization())
-        with self.sourceDirNode.child(self.customCssPath).open('wb') as cssfile:
-            cssfile.write(customStyles.encode())
-        with self.sourceDirNode.child(self.customJSPath).open('wb') as jsfile:
-            jsfile.write(self.defaultJSCustomization().encode())
-        self.page.site.shellCall('sphinx-build', self.sourceDirNode.internal_path, self.resultNode.internal_path, *args)
+        
+        self.processCssCustomizations()
+        self.processJsCustomizations()
+        self.processHtmlCustomizations()
+
+        args = [self.sourceDirNode.internal_path, self.resultNode.internal_path] + args
+        sphinx_build_main(args)
 
     def post_process(self):     
         with self.tblobj.recordToUpdate(self.handbook_id) as record:
@@ -149,7 +183,8 @@ class Main(BaseResourceBatch):
                 self.resultNode.move(self.publishedDocNode)
                 record['handbook_url'] = self.handbook_url
                 self.result_url = None
-        self.sphinxNode.delete()
+        if not self.db.application.getPreference('.save_src_debug',pkg='docu'):
+            self.sphinxNode.delete()
         self.db.commit()
 
         if self.db.application.getPreference('.manage_redirects',pkg='docu'):
@@ -221,10 +256,14 @@ class Main(BaseResourceBatch):
             rst = LINKFINDER.sub(self.fixLinks, rst)
 
             rst=rst.replace('[tr-off]','').replace('[tr-on]','')
+            footer= ''
             if record['author']:
-                footer = '\n.. sectionauthor:: %s\n'%record['author']
-            else:
-                footer= ''
+                footer = '\n.. sectionauthor:: %s\n' % (record['author'] or self.handbook_record['author'])
+            if self.db.application.getPreference('.handbooks_theme.last_update',pkg='docu'):
+                last_upd = translator.getTranslation('!!Publish date', language=self.handbook_record['language']).get('translation') or 'Publish date'
+                date_format = '%Y-%m-%d'if self.handbook_record['language'] == 'en' else '%d-%m-%Y' 
+                publish_date_str = record['publish_date'].strftime(date_format) if record['publish_date'] else ''
+                footer += f"""\n.. raw:: html\n\n   <p style="font-size:0.8em;">{last_upd} {publish_date_str}</p>"""
                 
             if n.attr['child_count']>0:
                 result.append('%s/%s.rst' % (name,name))
@@ -328,7 +367,6 @@ class Main(BaseResourceBatch):
         return '\n%s\n%s\n\n\n   %s' % (".. toctree::", '\n'.join(toc_options),'\n   '.join(elements))
 
     def invalidateCloudfrontCache(self):
-        import boto3, time
         client = boto3.client('cloudfront')
         response = client.create_invalidation(
                     DistributionId=self.db.application.getPreference('.cloudfront_distribution_id',pkg='docu'),
@@ -364,8 +402,6 @@ class Main(BaseResourceBatch):
     
     def table_script_parameters_pane(self,pane,**kwargs):   
         fb = pane.formbuilder(cols=1, border_spacing='5px')
-        #DP202112 Useful for local debugging 
-        #fb.checkbox(lbl='Skip images', value='^.skip_images')
         if self.db.application.getPreference('.manage_redirects',pkg='docu'):
             fb.checkbox(label='!![en]Skip redirects', value='^.skip_redirects')
         if self.db.application.getPreference('.cloudfront_distribution_id',pkg='docu'):
@@ -383,40 +419,80 @@ class Main(BaseResourceBatch):
             #                title='!![en]Notification recipients', 
             #                margin='2px', pbl_classes=True, addrow=False, delrow=False, height='200px')
 
+    def processCssCustomizations(self):
+        customCssPath='_static/custom.css' #DP Customizable?
+        cssStyles = [
+            s for s in [
+                self.db.application.getPreference('.base_css', pkg='docu'),
+                self.handbook_record.get('custom_styles'),
+                self.defaultCssCustomization()
+            ] if s
+        ]
+        with self.sourceDirNode.child(customCssPath).open('wb') as cssfile:
+            cssfile.write('\n'.join(cssStyles).encode())
+    
+    def processJsCustomizations(self):
+        customJSPath='_static/custom.js'    
+        jsStyles = [
+            s for s in [
+                self.db.application.getPreference('.base_js',pkg='docu'),
+                self.defaultJSCustomization()
+            ] if s
+        ]
+        with self.sourceDirNode.child(customJSPath).open('wb') as jsfile:
+            jsfile.write('\n'.join(jsStyles).encode())
+            
+    def processHtmlCustomizations(self):
+        extra_head = self.db.application.getPreference('.html_extra_head', pkg='docu')
+        if not extra_head:
+            return
+        layout_html = f"""
+            {{% extends "!layout.html" %}}
+
+            {{% block extrahead %}}
+              {{% raw %}}{{{{ super() }}}}{{% endraw %}}
+              {extra_head}
+            {{% endblock %}}
+            """
+        template_file = self.sourceDirNode.child('_templates', 'layout.html')
+        template_file.ensureParent()
+        with template_file.open('w') as f:
+            f.write(layout_html)
+            
     def defaultCssCustomization(self):
         return """/* override table width restrictions */
 
-@media screen and (min-width: 767px) {
+                @media screen and (min-width: 767px) {
 
-   .wy-table-responsive table td {
-      /* !important prevents the common CSS stylesheets from overriding
-         this as on RTD they are loaded after this stylesheet */
-      white-space: normal !important;
-   }
+                   .wy-table-responsive table td {
+                      /* !important prevents the common CSS stylesheets from overriding
+                         this as on RTD they are loaded after this stylesheet */
+                      white-space: normal !important;
+                   }
 
-   .wy-table-responsive {
-      overflow: visible !important;
-   }
-}
+                   .wy-table-responsive {
+                      overflow: visible !important;
+                   }
+                }
 
-.gnrexamplebox_title{
-    color:#2980B9;
-    cursor:pointer;
-}
-.gnrexamplebox_iframecont{
-    border:1px solid silver;
-    margin:5px;
-}
-"""
+                .gnrexamplebox_title{
+                    color:#2980B9;
+                    cursor:pointer;
+                }
+                .gnrexamplebox_iframecont{
+                    border:1px solid silver;
+                    margin:5px;
+                }
+                """
 
     def defaultJSCustomization(self):
         return """
-        var gnrExampleIframe = function(box,kw){
-            var src_root = window.location.port?window.location.origin+'/webpages/docu_examples':kw.examples_root;
-            var src = [src_root,kw.example_folder,kw.example_name].join('/');
-            src+=kw.parsstring;
-            var height = kw.height || '200px';
-            var width = kw.width || '100%'
-            box.innerHTML = '<div class="gnrexamplebox_iframecont"><iframe style="padding-bottom:3px; padding-right:3px; resize:vertical;" src="'+src+'" frameborder="0" height="'+height+'" width="'+width+'"></iframe></div>';
-        }
-    """
+                    var gnrExampleIframe = function(box,kw){
+                        var src_root = window.location.port?window.location.origin+'/webpages/docu_examples':kw.examples_root;
+                        var src = [src_root,kw.example_folder,kw.example_name].join('/');
+                        src+=kw.parsstring;
+                        var height = kw.height || '200px';
+                        var width = kw.width || '100%'
+                        box.innerHTML = '<div class="gnrexamplebox_iframecont"><iframe style="padding-bottom:3px; padding-right:3px; resize:vertical;" src="'+src+'" frameborder="0" height="'+height+'" width="'+width+'"></iframe></div>';
+                    }
+                """
