+++ projects/gnrcore/packages/sys/resources/services/storage/aws_s3.py
@@ -3,21 +3,20 @@
 #
 #  Copyright (c) 2013 Softwell. All rights reserved.
 
-
-from gnr.lib.services.storage import StorageService,StorageNode,StorageResolver
-from gnr.web.gnrbaseclasses import BaseComponent
-from gnr.core.gnrdecorator import public_method
-from gnr.core.gnrbag import Bag
-#from gnr.core.gnrlang import componentFactory
-import boto3
-import botocore
 import os
 import tempfile
 import mimetypes
 from datetime import datetime
 import warnings
+import traceback
 warnings.filterwarnings("ignore", category=ResourceWarning, message="unclosed.*<ssl.SSLSocket.*>")
 
+import boto3
+import botocore
+from smart_open import open as so_open
+
+from gnr.lib.services.storage import StorageService, StorageNode
+from gnr.web.gnrbaseclasses import BaseComponent
 
 class S3LocalFile(object):
     def __init__(self, mode='rb', bucket=None, key=None, s3_session=None):
@@ -58,7 +57,6 @@ class S3LocalFile(object):
                     result = self.file.__exit__(*exit_args)
                 self.file.close()
             except Exception as e:
-                import traceback
                 traceback.print_exc()
             finally:
                 os.unlink(self.name)
@@ -106,9 +104,12 @@ class S3TemporaryFilename(object):
 class Service(StorageService):
 
     def __init__(self, parent=None, bucket=None,
-        base_path=None, aws_access_key_id=None,
-        aws_secret_access_key=None, aws_session_token=None,
-        region_name=None, url_expiration=None, **kwargs):
+                base_path=None, aws_access_key_id=None,
+                aws_secret_access_key=None, aws_session_token=None,
+                region_name=None, url_expiration=None, write_in_local=None, 
+                readonly=None,versioned=True,
+                custom_endpoint=False, endpoint_url=None,
+                disable_cert_verify=False, **kwargs):
         self.parent = parent
         self.bucket = bucket
         self.base_path = (base_path or '').rstrip('/')
@@ -116,7 +117,20 @@ class Service(StorageService):
         self.aws_secret_access_key=aws_secret_access_key
         self.aws_session_token=aws_session_token
         self.region_name=region_name
+        self.endpoint_url = endpoint_url if custom_endpoint else None
+        self.disable_cert_verify = disable_cert_verify
         self.url_expiration = url_expiration or 3600
+        has_sys = 'gnrcore:sys' in self.parent.gnrapp.config['packages']
+        secondary = has_sys and self.parent.gnrapp.config['packages'].getAttr('gnrcore:sys').get('secondary')
+        local = getattr(parent,'_local_mode', False)
+        local_readonly = (local or secondary) and not write_in_local
+        self.readonly = readonly or local_readonly
+        self.versioned = versioned
+
+    @property
+    def is_versioned(self):
+        return self.versioned
+
     @property
     def location_identifier(self):
         return 's3/%s/%s' % (self.region_name, self.bucket)
@@ -148,6 +162,19 @@ class Service(StorageService):
             else:
                 raise
 
+
+    def versions(self,*args):
+        response =  self._client.list_object_versions(Bucket=self.bucket,Prefix=self.internal_path(*args))  
+        versions =  response.get('Versions',[])
+        result = {}
+        for v in versions:
+            replaced_v = result.get(v['ETag'])
+            if replaced_v and replaced_v.get('IsLatest'):
+                continue
+            result[v['ETag']] = v
+        return list(result.values())
+
+
     def md5hash(self,*args):
         bucket = self._head_object(*args)
         if bucket:
@@ -160,6 +187,8 @@ class Service(StorageService):
         pass
 
     def mkdir(self, *args, **kwargs):
+        if self.readonly:
+            return
         with self.open(*args+('.gnrdir',),mode='w') as dotfile:
             dotfile.write('.gnrdircontent')
 
@@ -202,9 +231,14 @@ class Service(StorageService):
             return
 
     def local_path(self, *args, **kwargs):
-        mode = kwargs.get('mode', 'r')
-        keep = kwargs.get('keep', False)
+        mode = kwargs.get('mode') or 'r'
+        keep = kwargs.get('keep') or False
         internalpath = self.internal_path(*args)
+        if self.readonly:
+            if 'b' in mode:
+                mode='rb'
+            else:
+                mode='r'
         return S3TemporaryFilename(bucket=self.bucket, key=internalpath,
             s3_session=self._session, mode=mode, keep=keep)
 
@@ -245,14 +279,23 @@ class Service(StorageService):
     @property
     def _client(self):
         if not hasattr(self, '_boto_client') or (hasattr(self,'_boto_client_ts') and (datetime.now()-self._boto_client_ts).seconds>120):
-            self._boto_client = self._session.client('s3', config= boto3.session.Config(signature_version='s3v4'))
+            session_kwargs={}
+            if self.endpoint_url:
+                session_kwargs['endpoint_url'] = self.endpoint_url
+            if self.disable_cert_verify:
+                session_kwargs['verify'] = False
+            self._boto_client = self._session.client('s3', config= boto3.session.Config(signature_version='s3v4'), **session_kwargs)
             self._boto_client_ts = datetime.now()
         return self._boto_client
 
     def delete_file(self, *args):
+        if self.readonly:
+            return
         self._client.delete_object(Bucket=self.bucket, Key=self.internal_path(*args))
 
     def delete_dir(self, *args):
+        if self.readonly:
+            return
         prefix = self.internal_path(*args)
         client = self._client
         prefix = prefix.strip('/')+'/' if prefix else ''
@@ -302,18 +345,28 @@ class Service(StorageService):
 
     def open(self, *args, **kwargs):
         kwargs['mode'] = kwargs.get('mode', 'rb')
-        from smart_open import open
-        open.DEFAULT_BUFFER_SIZE = 1024 * 1024
-        return open("s3://%s/%s"%(self.bucket,self.internal_path(*args)),
-            transport_params={'session':self._session}, **kwargs)
+        #version_id = kwargs.pop('version_id',None)
+        if self.readonly:
+            if 'b' in kwargs['mode']:
+                kwargs['mode'] = 'rb'
+            else:
+                kwargs['mode'] = 'r'
+        so_open.DEFAULT_BUFFER_SIZE = 1024 * 1024
+        version_id = kwargs.pop('version_id',None)
+        return so_open("s3://%s/%s"%(self.bucket,self.internal_path(*args)),
+            transport_params={'session':self._session, 'client': self._client,'version_id':version_id},**kwargs)
 
 
     def duplicateNode(self, sourceNode=None, destNode=None): # will work only in the same bucket
+        if self.readonly:
+            return
         self._s3_copy(source_bucket=sourceNode.service.bucket,
             source_key=sourceNode.internal_path,
             dest_bucket=destNode.service.bucket, dest_key=destNode.internal_path)
 
     def renameNode(self, sourceNode=None, destNode=None):
+        if self.readonly:
+            return
         self._s3_copy(source_bucket=sourceNode.service.bucket,
             source_key=sourceNode.internal_path,
             dest_bucket=destNode.service.bucket, dest_key=destNode.internal_path)
@@ -365,11 +418,16 @@ class ServiceParameters(BaseComponent):
     def service_parameters(self,pane,datapath=None,**kwargs):
         bc = pane.borderContainer()
         fb = bc.contentPane(region='top').formbuilder(datapath=datapath)
+        fb.checkbox(value='^.readonly',lbl='Read Only')
+        fb.checkbox(value='^.write_in_local',lbl='Write enabled on local/secondary machine')
         fb.textbox(value='^.bucket',lbl='Bucket')
         fb.textbox(value='^.base_path',lbl='Base path')
         fb.textbox(value='^.aws_access_key_id',lbl='Aws Access Key Id')
         fb.textbox(value='^.aws_secret_access_key',lbl='Aws Secret Access Key')
         fb.textbox(value='^.region_name',lbl='Region Name')
+        fb.checkbox(value='^.custom_endpoint',lbl='Custom Endpoint')
+        fb.textbox(value='^.endpoint_url',lbl='Endpoint Url', hidden='==!custom', custom='^.custom_endpoint')
+        fb.checkbox(value='^.disable_cert_verify',lbl='Disable Certificate Verification')
         bc.storageTreeFrame(frameCode='bucketStorage',storagepath='^#FORM.record.service_name?=#v+":"',
                                 border='1px solid silver',margin='2px',rounded=4,
                                 region='center',preview_region='right',
